spiflash: correctly output WREN missing warnings
Original code would try and output an annotation with es well in the
past, resulting in a sliver that couldn't be zoomed in on.

Instead, correctly update the end of the command as is done with other
commands, and wait to output the warning until we've finished with the
command entirely.

Signed-off-by: Karl Palsson <karl.palsson@marel.com>