spiflash: check RDID against provided
This overhauls the data tables to make them a little bit easier to work
with in the future, by using classes, instead of big string dicts.

The old code was a little mixed between showing the RDID results, or
showing what had been selected, so make that clear, and use warnings if
things aren't as we expect.  We also warn if the selected part doesn't
provide IDs, but we saw one anyway.  Decoding _may_ be just fine, but it
may be wrong.  Checked also with sigrok-dumps for parts without rdid.

For "manufacturer" id, we've gone with keeping tables in the style of
memtest86, using an upper byte for the page number, rather than trying
to keep literals like "0x7f7f7f7f7f7fc2" or similar ala flashrom.

TODO: this does _not_ overhaul the related rems/res/rdp style ID
decoding, but that really needs similar treatment.

Signed-off-by: Karl Palsson <karl.palsson@marel.com>